4-面向交叉学科的Python程序设计/levelG/code/requirement.txt
brotlipy==0.7.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi @ file:///home/builder/ci_310/cffi_1642753365720/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///tmp/build/80754af9/click_1646056706450/work
cryptography @ file:///tmp/build/80754af9/cryptography_1652101588599/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython==0.29.30
filelock @ file:///opt/conda/conda-bld/filelock_1647002191454/work
fonttools==4.25.0
future @ file:///home/builder/ci_310/future_1640790123501/work
huggingface-hub @ file:///tmp/build/80754af9/huggingface_hub_1639662742275/work
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imgkit==1.2.2
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648562408262/work
jieba==0.42.1
Jinja2==3.1.2
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
kiwisolver @ file:///opt/conda/conda-bld/kiwisolver_1653292039266/work
lxml==4.9.0
MarkupSafe==2.1.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1647441664166/work
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
munkres==1.1.4
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1652801679809/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.3
patsy==0.5.2
Pillow==9.0.1
prettytable==3.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyecharts==1.9.1
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt5-stubs==5.15.6.0
PyQtWebEngine-Qt5==5.15.2
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2022.1
PyYAML==6.0
regex @ file:///tmp/build/80754af9/regex_1648429580840/work
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
sacremoses @ file:///tmp/build/80754af9/sacremoses_1633107328213/work
scikit-learn==1.1.1
scipy==1.8.1
simplejson==3.17.6
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
statsmodels==0.13.2
threadpoolctl==3.1.0
tokenizers @ file:///opt/conda/conda-bld/tokenizers_1651822590771/work
torch==1.10.2
tornado @ file:///home/builder/ci_310/tornado_1640795392144/work
tqdm @ file:///opt/conda/conda-bld/tqdm_1650891076910/work
transformers @ file:///opt/conda/conda-bld/transformers_1651834311006/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1650637206367/work
wcwidth==0.2.5
zipp @ file:///opt/conda/conda-bld/zipp_1652341764480/work
